LayoutTests/fullscreen/full-screen-enter-while-exiting-multiple-elements-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Fullscreen request aborted by a fullscreen request for another element.
Enter fullscreen with target1
RUN(target1.requestFullscreen())
EVENT(fullscreenchange)
Enter fullscreen with target2
RUN(target2.requestFullscreen())
EVENT(fullscreenchange)
Exit fullscreen with target1
RUN(document.exitFullscreen())
Attempt to enter fullscreen with target3
RUN(target3.requestFullscreen())
EVENT(fullscreenchange)
EVENT(fullscreenerror)
Entering fullscreen with target3
RUN(target3.requestFullscreen())
EVENT(fullscreenchange)
END OF TEST
